refactor: dkim registry logic abstracted away
