+++ contracts/Nargo.toml
@@ -2,6 +2,6 @@
 members = [
     "address_note",
     "generators",
-    "z_imburse_escrow",
+    # "z_imburse_escrow",
     "z_imburse_registry",
 ]
