+++ contracts/z_imburse_registry/src/library_methods/mod.nr
@@ -1,0 +1,1 @@
+mod dkim;
