+++ src/artifacts/contracts/ZImburseEscrow.ts
@@ -33,37 +33,19 @@ import {
   type Wallet,
   type WrappedFieldLike,
 } from '@aztec/aztec.js';
-import ZImburseEscrowContractArtifactJson from './ZImburseEscrow.json' assert { type: 'json' };
-//@ts-ignore
-//@ts-ignore
-//@ts-ignore
-//@ts-ignore
-//@ts-ignore
-//@ts-ignore
-//@ts-ignore
-//@ts-ignore
-//@ts-ignore
-//@ts-ignore
-//@ts-ignore
-//@ts-ignore
-//@ts-ignore
-//@ts-ignore
-//@ts-ignore
-//@ts-ignore
-//@ts-ignore
-//@ts-ignore
+import ZImburseEscrowContractArtifactJson from './z_imburse_escrow-ZImburseEscrow.json' assert { type: 'json' };
 export const ZImburseEscrowContractArtifact = loadContractArtifact(ZImburseEscrowContractArtifactJson as NoirCompiledContract);
 
+
+      export type EntitlementNullified = {
+        randomness: FieldLike
+      }
+    
 
       export type SpotReimbursementClaimed = {
         claimant: AztecAddressLike
 amount: FieldLike
 verifier_id: (bigint | number)
-      }
-    
-
-      export type EntitlementNullified = {
-        randomness: FieldLike
       }
     
 
@@ -156,7 +138,7 @@ entitlements: {
     }
     
 
-  public static get notes(): ContractNotes<'AddressNote' | 'TransparentNote' | 'TokenNote' | 'EntitlementNote'> {
+  public static get notes(): ContractNotes<'AddressNote' | 'TransparentNote' | 'UintNote' | 'EntitlementNote'> {
     return {
       AddressNote: {
           id: new NoteSelector(2232136525),
@@ -164,13 +146,13 @@ entitlements: {
 TransparentNote: {
           id: new NoteSelector(3193649735),
         },
-TokenNote: {
-          id: new NoteSelector(2350566847),
+UintNote: {
+          id: new NoteSelector(202136239),
         },
 EntitlementNote: {
           id: new NoteSelector(4112046478),
         }
-    } as ContractNotes<'AddressNote' | 'TransparentNote' | 'TokenNote' | 'EntitlementNote'>;
+    } as ContractNotes<'AddressNote' | 'TransparentNote' | 'UintNote' | 'EntitlementNote'>;
   }
   
 
@@ -244,9 +226,25 @@ EntitlementNote: {
       };
     }
 
-    public static get events(): { SpotReimbursementClaimed: {decode: (payload: L1EventPayload | UnencryptedL2Log | undefined) => SpotReimbursementClaimed | undefined, eventSelector: EventSelector, fieldNames: string[] }, EntitlementNullified: {decode: (payload: L1EventPayload | UnencryptedL2Log | undefined) => EntitlementNullified | undefined, eventSelector: EventSelector, fieldNames: string[] }, RecurringReimbursementClaimed: {decode: (payload: L1EventPayload | UnencryptedL2Log | undefined) => RecurringReimbursementClaimed | undefined, eventSelector: EventSelector, fieldNames: string[] } } {
+    public static get events(): { EntitlementNullified: {decode: (payload: L1EventPayload | UnencryptedL2Log | undefined) => EntitlementNullified | undefined, eventSelector: EventSelector, fieldNames: string[] }, SpotReimbursementClaimed: {decode: (payload: L1EventPayload | UnencryptedL2Log | undefined) => SpotReimbursementClaimed | undefined, eventSelector: EventSelector, fieldNames: string[] }, RecurringReimbursementClaimed: {decode: (payload: L1EventPayload | UnencryptedL2Log | undefined) => RecurringReimbursementClaimed | undefined, eventSelector: EventSelector, fieldNames: string[] } } {
     return {
-      SpotReimbursementClaimed: {
+      EntitlementNullified: {
+        decode: this.decodeEvent(EventSelector.fromSignature('EntitlementNullified(Field)'), {
+    "fields": [
+        {
+            "name": "randomness",
+            "type": {
+                "kind": "field"
+            }
+        }
+    ],
+    "kind": "struct",
+    "path": "ZImburseEscrow::EntitlementNullified"
+}),
+        eventSelector: EventSelector.fromSignature('EntitlementNullified(Field)'),
+        fieldNames: ["randomness"],
+      },
+SpotReimbursementClaimed: {
         decode: this.decodeEvent(EventSelector.fromSignature('SpotReimbursementClaimed((Field),Field,u8)'), {
     "fields": [
         {
@@ -284,22 +282,6 @@ EntitlementNote: {
 }),
         eventSelector: EventSelector.fromSignature('SpotReimbursementClaimed((Field),Field,u8)'),
         fieldNames: ["claimant","amount","verifier_id"],
-      },
-EntitlementNullified: {
-        decode: this.decodeEvent(EventSelector.fromSignature('EntitlementNullified(Field)'), {
-    "fields": [
-        {
-            "name": "randomness",
-            "type": {
-                "kind": "field"
-            }
-        }
-    ],
-    "kind": "struct",
-    "path": "ZImburseEscrow::EntitlementNullified"
-}),
-        eventSelector: EventSelector.fromSignature('EntitlementNullified(Field)'),
-        fieldNames: ["randomness"],
       },
 RecurringReimbursementClaimed: {
         decode: this.decodeEvent(EventSelector.fromSignature('RecurringReimbursementClaimed((Field),Field,u8,Field)'), {
